requirements.txt
azure-common==1.1.28
azure-core==1.29.5
azure-cosmos==4.5.1
azure-keyvault-secrets==4.7.0
azure-mgmt-compute==30.3.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azure-search-documents==11.3.0
azure-storage-blob==12.19.0
bidict==0.22.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
eventlet==0.33.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
requests-oauthlib==1.3.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
